+++ ImageFormatAggregator.cpp
@@ -35,7 +35,7 @@ HGLOBAL CImageFormatAggregator::GetHGlobal()
 		return NULL;
 	}
 
-	int count = m_images.GetCount();
+	int count = (int)m_images.GetCount();
 	for (int i = 0; i < count; i++)
 	{
 		CClipFormat clip = m_images[i];
